move process_cache
